# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Robot Shop",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/admin/roles": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "搜索时支持角色名和 Casbin Role 筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "获取角色列表",
                "operationId": "ListRoles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "角色名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Casbin Role",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.RoleSearchResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一个新的角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "创建角色",
                "operationId": "RoleCreate",
                "parameters": [
                    {
                        "description": "角色数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.RoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/admin/roles/permission": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取指定角色的权限列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "获取角色权限",
                "operationId": "GetRolePermission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "角色名",
                        "name": "role",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.GetRolePermissionResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新指定角色的权限列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "更新角色权限",
                "operationId": "UpdateRolePermission",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.UpdateRolePermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/admin/roles/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "目前只允许更新角色名称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "更新角色",
                "operationId": "RoleUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "角色数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.RoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定ID的角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "删除角色",
                "operationId": "RoleDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名",
                    "type": "string"
                },
                "role": {
                    "description": "Casbin Role",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "RoleList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Role"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "backend_api_v1.GetRolePermissionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/backend_api_v1.GetRolePermissionResponseData"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.GetRolePermissionResponseData": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
        "backend_api_v1.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_api_v1.RoleRequest": {
            "type": "object",
            "required": [
                "name",
                "role"
            ],
            "properties": {
                "name": {
                    "description": "角色名",
                    "type": "string",
                    "example": "admin"
                },
                "role": {
                    "description": "Casbin Role",
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "backend_api_v1.RoleSearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/RoleList"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.UpdateRolePermissionRequest": {
            "type": "object",
            "required": [
                "list",
                "role"
            ],
            "properties": {
                "list": {
                    "description": "权限列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        ""
                    ]
                },
                "role": {
                    "description": "角色名",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}